% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse}
\alias{analyse}
\title{Data analyse and antenna parameters estimation}
\usage{
analyse()
}
\value{
parameters like N, L, C for antenna design
}
\description{
Data analyse and antenna parameters estimation
}
\examples{
analyse()

}
