% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfid.R
\name{Best}
\alias{Best}
\title{Magnetic field estimation. Evaluates the magnetic field as a function of the antenna current}
\usage{
Best(I)
}
\arguments{
\item{I}{current estimated from In function}
}
\description{
Magnetic field estimation. Evaluates the magnetic field as a function of the antenna current
}
