% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildEstimator.R
\name{buildEstimator}
\alias{buildEstimator}
\title{Builds an estimator based on the matrix of regression coefficients depending on a parameter value}
\usage{
buildEstimator(matrice, param)
}
\arguments{
\item{matrice}{matrix of regression coefficients}

\item{param}{value of the parameter. ex: F=125. The frequencies are indicated in kHz}
}
\value{
a string representing the estimator for the parameter value given as parameter
}
\description{
Builds an estimator based on the matrix of regression coefficients depending on a parameter value
}
\examples{
buildEstimator(CS[3], 125)

}
