% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfid.R
\name{Cresonnance}
\alias{Cresonnance}
\title{Resonance frequency in kHz for L in Henry}
\usage{
Cresonnance(L, F)
}
\arguments{
\item{L}{inductance of the antenna}
}
\description{
Resonance frequency in kHz for L in Henry
}
