% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateurs.R
\name{estimateurs}
\alias{estimateurs}
\title{Fourni la liste des fonctions de régression d'un groupe pour un ensemble de modèles
l'appel se fait sur un groupe: LN, RN, LF, RF}
\usage{
estimateurs(models, groupe)
}
\arguments{
\item{models}{models possède 4 groupes:LN, RN, LF, RF}

\item{groupe}{4 groupes:LN, RN, LF, RF}
}
\value{
a liste of models
}
\description{
Fourni la liste des fonctions de régression d'un groupe pour un ensemble de modèles
l'appel se fait sur un groupe: LN, RN, LF, RF
}
\examples{
M<-getModels(TAB)
estimateurs(M, "LN")
}
