% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{estimators}
\alias{estimators}
\title{Provide a list of linear models for a group of data: LN, RN, LF, RF
To be call based on a group}
\usage{
estimators(models, groupe)
}
\arguments{
\item{models}{made of 4 groups:LN, RN, LF, RF}

\item{groupe}{4 groups:LN, RN, LF, RF}
}
\value{
a list of models: LN, RN, LF, LF
}
\description{
Provide a list of linear models for a group of data: LN, RN, LF, RF
To be call based on a group
}
\examples{
M<-getModels(TAB)
est<-estimateurs(M, "LN")
}
