% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalEstimator.R
\name{evalEstimator}
\alias{evalEstimator}
\title{Exécute le calcul des valeurs de la grandeur d'interêt suivant l'estimateur}
\usage{
evalEstimator(estimator, param, interval, titre = "", labX = "", labY = "")
}
\arguments{
\item{estimator}{estimateur. Polynôme en forme de string}

\item{param}{valeur du paramètre. ex: F=125. Les fréquences sont indiquées en kHz}

\item{interval}{étendue de l'interval de calcul. ex: 60:120 ou seq(60:120, by=0.1) pour le nbre de spires si le paramètre est F}

\item{titre}{titre du graphe produit - optionnel}

\item{labX}{label axe x - optionnel}

\item{labY}{label axe y - optionnel}
}
\value{
un numeric
}
\description{
Exécute le calcul des valeurs de la grandeur d'interêt suivant l'estimateur
}
