% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalEstimator2.R
\name{evalEstimator2}
\alias{evalEstimator2}
\title{Executes the calculation of the values of the quantity of interest according to the estimator}
\usage{
evalEstimator2(matrice, param, interval)
}
\arguments{
\item{matrice}{matrix of regression coefficients}

\item{param}{value of the parameter. ex: F=125. The frequencies are indicated in kHz}

\item{interval}{range of the calculation interval. ex: 60:120 or seq(60:120, by=0.1) for the number of turns if the parameter is F}
}
\value{
a list of values according the estimator on the whole range of interval calculation
}
\description{
Executes the calculation of the values of the quantity of interest according to the estimator
}
\examples{
evalEstimator2(CS[3], 125, c(60, 120))

}
