% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extModels.R
\name{extModels}
\alias{extModels}
\title{Extrapoles the value of the inductance or resistance for a given frequency based on the measurements recorded}
\usage{
extModels(data, parameter, inductance, plots)
}
\arguments{
\item{data}{data frame representing data}

\item{parameter}{the frequency in kHz used for extrapolate the models of inductance or resistance antenna}

\item{inductance}{boolean. If TRUE the inductance for the parameter value is estimated, if FALSE, the resistance is estimated.}

\item{plots}{boolean. Enable plots}
}
\value{
the coefficients of the extrapolated model for the antenna inductance or the resistance
}
\description{
Extrapoles the value of the inductance or resistance for a given frequency based on the measurements recorded
}
\examples{

TAB<-getMeasures("ant046")
CO<-extModels(TAB, 125, FALSE, TRUE)
CO<-extModels(TAB, 125, TRUE, TRUE)

}
