% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfid.R
\name{Fres}
\alias{Fres}
\title{Estimation of tuning frequency as a function of circuit parameters}
\usage{
Fres(L, C)
}
\arguments{
\item{L}{Inductance of the antenna}

\item{C}{Capacitance in pf}
}
\description{
Estimation of tuning frequency as a function of circuit parameters
}
