% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllGraphs.R
\name{getAllGraphs}
\alias{getAllGraphs}
\title{Provides the set of cumulative graphs with insertion of the simulated curve}
\usage{
getAllGraphs(params)
}
\arguments{
\item{params}{list of parameter values}
}
\value{
plots of quantities of interest
}
\description{
Provides the set of cumulative graphs with insertion of the simulated curve
}
\examples{
getAllGraphs(c(110, 110, 125, 90))
 getAllGraphs(c(N1, N2, F1, F2))
}
