% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{getC2}
\alias{getC2}
\title{Equivalent capacitance value of the antenna divider voltage getter}
\usage{
getC2()
}
\value{
a numeric
}
\description{
Equivalent capacitance value of the antenna divider voltage getter
}
