% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimator.R
\name{getEstimator}
\alias{getEstimator}
\title{Generates the estimator from the regression coefficients of the model parameters}
\usage{
getEstimator(matrice)
}
\arguments{
\item{matrice}{matrix list of regression coefficients on the coefficients extracted from the measurement data}
}
\value{
Returns the equation Y(x) representing the quantity of interest L or R as a function of x (N or F)
Y(y) <-C(x) + B(x) * y + A(x) * y^2
A, B, C factor in the form of polymorphs of degree 2, result of the regression on the parameters
of the models
y: variable of interest: N or F
x: parameter: N for the xF group; F for the xN group
the coefficients of Y depend on x. Each coefficient is a polynomial of degree 2
}
\description{
Generates the estimator from the regression coefficients of the model parameters
}
\examples{
CS<-regMods("data")
matrice= CS[1]

}
