% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfid.R
\name{getMaxBest}
\alias{getMaxBest}
\title{Evaluates the number of turns for peak value of the antenna magnetic field}
\usage{
getMaxBest(matrice, param, interval)
}
\arguments{
\item{matrice}{Matrix of coefficents for antenna resistance estimator}

\item{param}{the value of the frequency for evaluation of the resistance antenna estimator}

\item{interval}{The range of value to estimate the peak value of the antenna magnetic field}
}
\value{
the number of turn to achieve to peak value of the antenna magnetic field
}
\description{
Evaluates the number of turns for peak value of the antenna magnetic field
}
