% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfid.R
\name{getMaxIest}
\alias{getMaxIest}
\title{Evaluates the number of turns for peak value of the antenna current}
\usage{
getMaxIest(matrice, param, interval)
}
\arguments{
\item{matrice}{Matrix of coefficents for antenna resistance estimator}

\item{param}{the value of the frequency for evaluation of the resistance antenna estimator}

\item{interval}{The range of value to estimate the peak value of the antenna current}
}
\value{
the number of turn to achieve to peak value of the antenna current
}
\description{
Evaluates the number of turns for peak value of the antenna current
}
