% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModParams.R
\name{getModparams}
\alias{getModparams}
\title{lists of model parameters for a group: RN, LN, LF, RF  contained in a set designated by models}
\usage{
getModparams(models, groupe)
}
\arguments{
\item{models}{List of models provided by getmModels()}

\item{groupe}{groupe of models: LN, RN, LF, RF}
}
\description{
lists of model parameters for a group: RN, LN, LF, RF  contained in a set designated by models
}
\examples{

M<-getModels(TAB)
getModparams(M, "LN")

}
