% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfid.R
\name{getVant}
\alias{getVant}
\title{Provide antenna voltage given the number of turns, the resistance and inductance value of the antenna}
\usage{
getVant(N, R, L)
}
\arguments{
\item{N}{number of turns}

\item{R}{antenna resistance in Ohm}

\item{L}{antenna inductance in Henry}
}
\description{
Provide antenna voltage given the number of turns, the resistance and inductance value of the antenna
}
\examples{
v<-getVant(70, 200, 1.6e-3)

}
