% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfid.R
\name{In}
\alias{In}
\title{Evaluates the antenna current from the antenna resistance estimator}
\usage{
In(N, Restimator)
}
\arguments{
\item{N}{number of turns of the antenna}

\item{Restimator}{antenna resistance estimator obtained from the regression coefficients and given in string form}
}
\description{
Evaluates the antenna current from the antenna resistance estimator
}
