% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfid.R
\name{Lattendue}
\alias{Lattendue}
\title{Evaluates the inductance in mH for the resonance at the frequency F given in kHz and the capacitance C given in pF}
\usage{
Lattendue(C, F)
}
\arguments{
\item{C}{capacitance value in pF}
}
\description{
Evaluates the inductance in mH for the resonance at the frequency F given in kHz and the capacitance C given in pF
}
\examples{
Lattendue(980e-12, 125)
[1] 1.654223

}
