% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisEst.R
\name{optimisEst}
\alias{optimisEst}
\title{Returns the value of the abcissa of the N or F estimator which provides the value of the quantity
of interest: L or R}
\usage{
optimisEst(coefs, value, interval)
}
\arguments{
\item{coefs}{estimated parameters of the model}

\item{value}{expected value of the quantity of interest}
}
\description{
Returns the value of the abcissa of the N or F estimator which provides the value of the quantity
of interest: L or R
}
\examples{
m<-optimisEst(CS[4], 2000, 125, c(60, 120)

}
