% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paraModsRegs.R
\name{paraModsRegs}
\alias{paraModsRegs}
\title{Deduce models on the parameters from a matrix of parameters}
\usage{
paraModsRegs(matrice, abscisse)
}
\arguments{
\item{matrice}{matrix from a group RN, RL, LN, or LF via getMatParams()}

\item{abscisse}{vector of the values of frequencies F or number of turns N which were used for the measurements
abscisse is the parameter of the studied curve network: F for RN, LN; N for RF, LF}
}
\description{
Deduce models on the parameters from a matrix of parameters
}
