% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunction.R
\name{plotFunction}
\alias{plotFunction}
\title{Plots a function}
\usage{
plotFunction(func, xm, xM, main, xlab, ylab)
}
\arguments{
\item{func}{the function to plot given in string form}

\item{xm}{minimum value of the interval of representatiion}

\item{xM}{maximumvalue of the interval of representatiion}

\item{main}{titre of the plot}

\item{xlab}{label of axis x}

\item{ylab}{label of axis y}
}
\description{
Plots a function
}
\examples{
f<-"x^2 + 2*x +3"
plotFunction(f, 0, 20, "title", "axis x name", "axis y name")

}
