% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGroups.R
\name{plotGroups}
\alias{plotGroups}
\title{graph of joint representations parametrized in F or N}
\usage{
plotGroups(data, groupe, estimation, param)
}
\arguments{
\item{data}{dataset from getMeasures()}

\item{groupe}{group of models: "RN", "LN", "RF", "LF"}

\item{estimation}{data from evalEstimator2}

\item{param}{value of the parameter in the models. Ex: 125kHz}
}
\description{
graph of joint representations parametrized in F or N
}
