% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfid.R
\name{plotI}
\alias{plotI}
\title{Plot the antenna current vs Turns}
\usage{
plotI(TAB)
}
\arguments{
\item{TAB}{data frame representing the data}
}
\description{
Plot the antenna current vs Turns
}
\examples{
TAB<-getMeasures("ant046")
plotI(TAB)

}
