% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMODSParams.R
\name{plotMODSParams}
\alias{plotMODSParams}
\title{Plots the regression curves for the model coefficients: RN, LN, RF, or LF}
\usage{
plotMODSParams(matrice, intervalle)
}
\arguments{
\item{matrice}{matrix list of regression coefficients on the coefficients extracted from the measurement data}

\item{intervalle}{vectors of the measurement interval. Related to measurement points: frequencies or nTypes.
e.g:intervalle=c(60,120) or intervalle=c(10, 100)
e.g:matrice= CS[1]
matrice= CS[1]}
}
\description{
Plots the regression curves for the model coefficients: RN, LN, RF, or LF
}
\examples{
CS<-regMods("data")
}
