% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regMods.R
\name{regMods}
\alias{regMods}
\title{Returns a list of all group coefficient matrices: RN, RF, LN, LF
after regression on all the coefficients matrices}
\usage{
regMods(data)
}
\arguments{
\item{data}{data file to be processed}
}
\description{
Returns a list of all group coefficient matrices: RN, RF, LN, LF
after regression on all the coefficients matrices
}
\examples{
regMods("data")

}
