% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfid.R
\name{RforI}
\alias{RforI}
\title{Optimise the value of the antenna resistance given a current antenna value}
\usage{
RforI(I)
}
\arguments{
\item{I}{}
}
\description{
Optimise the value of the antenna resistance given a current antenna value
}
