% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveGraphPng.R
\name{saveGraphPng}
\alias{saveGraphPng}
\title{Recording a graph}
\usage{
saveGraphPng(fileName, p)
}
\arguments{
\item{fileName}{Nom of the plot file}

\item{p}{plot to save}
}
\description{
Recording a graph
}
