% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{setDataSource}
\alias{setDataSource}
\title{Set source folder for loading data}
\usage{
setDataSource(source)
}
\arguments{
\item{name}{of the source folder as string}
}
\value{
nothing
}
\description{
Set source folder for loading data
}
\examples{
setDataSource("folder")

}
