% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{setR}
\alias{setR}
\title{Radius of antenna setter. Use for magnetic field estimation}
\usage{
setR(r)
}
\arguments{
\item{r}{}
}
\value{
nothing
}
\description{
Radius of antenna setter. Use for magnetic field estimation
}
