% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{setRefFrequency}
\alias{setRefFrequency}
\title{Set the reference frequency for estimations. Usely this parameter is set as 125
as 125kHz}
\usage{
setRefFrequency(freq)
}
\arguments{
\item{freq}{The reference frequency in kHz}
}
\value{
nothing
}
\description{
Set the reference frequency for estimations. Usely this parameter is set as 125
as 125kHz
}
\examples{
setRefFrequency(125)

}
