% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{setRser}
\alias{setRser}
\title{Serial antenna resistance setter}
\usage{
setRser(r)
}
\arguments{
\item{r}{}
}
\value{
nothing
}
\description{
Serial antenna resistance setter
}
\examples{
setRser(18)

}
