% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{setVdd}
\alias{setVdd}
\title{Voltage of the antenna driver setter}
\usage{
setVdd(v)
}
\arguments{
\item{v}{}
}
\value{
nothing
}
\description{
Voltage of the antenna driver setter
}
\examples{
setVdd(10)
}
