% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{setZ}
\alias{setZ}
\title{Setter for distance reference for magnetic field estimation}
\usage{
setZ(z)
}
\arguments{
\item{z}{}
}
\value{
nothig
}
\description{
Setter for distance reference for magnetic field estimation
}
