% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{summaryEnv}
\alias{summaryEnv}
\title{Summary of the constants use for all estimations in the package}
\usage{
summaryEnv()
}
\value{
a list of constants
}
\description{
Summary of the constants use for all estimations in the package
}
