% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{updateC2}
\alias{updateC2}
\title{Equivalent capacitance value of the antenna divider voltage}
\usage{
updateC2()
}
\value{
a numeric
}
\description{
Equivalent capacitance value of the antenna divider voltage
}
