% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfid.R
\name{Vant}
\alias{Vant}
\title{Antenna voltage estimation}
\usage{
Vant(N, R, L)
}
\arguments{
\item{N}{number of turns of the antenna}

\item{R}{resistance antenna}

\item{L}{inductance antenna}
}
\description{
Antenna voltage estimation
}
