% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse}
\alias{analyse}
\title{Data analyse and antenna parameters estimation.Provide N tours according an inductance value or a antenna current selected.
Provide resonant capacitance according an inductance value estimated.}
\usage{
analyse(data)
}
\arguments{
\item{data}{The data file to be processed. The must be save with csv extention.}
}
\value{
parameters like N, L, C for antenna design
}
\description{
Data analyse and antenna parameters estimation.Provide N tours according an inductance value or a antenna current selected.
Provide resonant capacitance according an inductance value estimated.
}
\examples{
analyse("data")

}
