% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfid.R
\name{Best}
\alias{Best}
\title{Evaluates the antenna magnetic field based on the antenna resistance estimator and the number of turns}
\usage{
Best(N, Restimator)
}
\arguments{
\item{N}{number of turns}

\item{Restimator}{an estiamtor of the antenna resistance obtained from the regression coefficients and given in string form}
}
\description{
Evaluates the antenna magnetic field based on the antenna resistance estimator and the number of turns
}
