% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfid.R
\name{Cres}
\alias{Cres}
\title{Estimation of the resonance capacity}
\usage{
Cres(N, L)
}
\arguments{
\item{N}{number of turns of the antenna}

\item{L}{inductance of the antenna}
}
\description{
Estimation of the resonance capacity
}
