% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfid.R
\name{Cresonnance}
\alias{Cresonnance}
\title{Resonnance capacitance calculation}
\usage{
Cresonnance(L, F)
}
\arguments{
\item{L}{inductance of the antenna in mH}

\item{F}{frequency in kHz}
}
\description{
Resonnance capacitance calculation
}
\examples{
Cresonnance(1.65, 125)
[1] 9.444968e-10

}
