% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalEstimator2.R
\name{evalEstimator2}
\alias{evalEstimator2}
\title{Executes the calculation of the values of the quantity of interest according to the estimator coefficients}
\usage{
evalEstimator2(coefs, parameter)
}
\arguments{
\item{coefs}{estimated coefficients of the model}

\item{parameter}{the value to provide the quantity of interest value}
}
\value{
a numeric
}
\description{
Executes the calculation of the values of the quantity of interest according to the estimator coefficients
}
\examples{
evalEstimator2(COEFSL, 524)

}
