% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllGraphs.R
\name{getAllGraphs}
\alias{getAllGraphs}
\title{Provides the set of cumulative graphs with insertion of the simulated curve}
\usage{
getAllGraphs(file, params)
}
\arguments{
\item{file}{name of the file to be processed given as a string}

\item{params}{list of parameter values}
}
\value{
plots of quantities of interest
}
\description{
Provides the set of cumulative graphs with insertion of the simulated curve
}
\examples{
getAllGraphs("data", c(110, 110, 125, 90))
getAllGraphs("data", c(N1, N2, F1, F2))
}
