% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllGraphs.R
\name{getAllGraphs}
\alias{getAllGraphs}
\title{Provides two sets of cumulative graphs with insertion of the simulated curve}
\usage{
getAllGraphs(data, param)
}
\arguments{
\item{data}{data frame representing the measurements to analyse provided by the function getMeasures based on the data file with csv extention}

\item{param}{the frequency value in kHz for extrapolations.}
}
\value{
plots of quantities of interest
}
\description{
Provides two sets of cumulative graphs with insertion of the simulated curve
}
\examples{
TAB<-getMeasures("ant046")
getAllGraphs(TAB, 125)
}
