% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllGraphs.R
\name{getAllGraphs}
\alias{getAllGraphs}
\title{Provides two sets of cumulative graphs with insertion of the simulated curve}
\usage{
getAllGraphs(data, param)
}
\arguments{
\item{params}{value of the frequency estimation in kHz, e.g: 125}

\item{file}{name of the file to be processed given as a string. The file must have csv extention}
}
\value{
plots of quantities of interest
}
\description{
Provides two sets of cumulative graphs with insertion of the simulated curve
}
\examples{
TAB<-getMeasures("ant046")
getAllGraphs(TAB, 125)
}
