% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeasures.R
\name{getMeasures}
\alias{getMeasures}
\title{Loads data from file. The file must be located on the appropriate folder. The folder must be named data and placed at the root
of the project folder}
\usage{
getMeasures(file, s = ";", d = ",")
}
\arguments{
\item{file}{Name of the data file. The file must be saved with .csv extention}

\item{s}{The separator of data in the data file: "," or ";"}

\item{d}{The decimal point used in the data file: "," or "."}
}
\value{
A data frame made of the data recorded
}
\description{
Loads data from file. The file must be located on the appropriate folder. The folder must be named data and placed at the root
of the project folder
}
\examples{
TAB<-getMeasures("data", ",", ".")
}
