% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfid.R
\docType{data}
\name{In}
\alias{In}
\title{Evaluates the antenna current from the antenna resistance estimator}
\format{
An object of class \code{numeric} of length 9.
}
\usage{
In
}
\arguments{
\item{N}{number of turns of the antenna}

\item{Restimator}{antenna resistance estimator obtained from the regression coefficients and given in string form}
}
\description{
Evaluates the antenna current from the antenna resistance estimator
}
\keyword{datasets}
