% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfid.R
\name{RforI}
\alias{RforI}
\title{Optimises the value of the antenna resistance given a current antenna value}
\usage{
RforI(I)
}
\arguments{
\item{I}{expected antenna current in mA}
}
\description{
Optimises the value of the antenna resistance given a current antenna value
}
